requirements.txt
channels
django-channels-presence
